getContaining methods refactoring (#648)
* move getContainingFunction method to utils

* create method to get the containing SourceUnit of a node
