+++ src/ast/ast.ts
@@ -24,7 +24,7 @@ import { printNode } from '../utils/astPrinter';
 import { TranspileFailedError } from '../utils/errors';
 import { Implicits } from '../utils/implicits';
 import { createBlock } from '../utils/nodeTemplates';
-import { isExternalCall, mergeImports } from '../utils/utils';
+import { getContainingSourceUnit, isExternalCall, mergeImports } from '../utils/utils';
 import { CairoFunctionDefinition } from './cairoNodes';
 
 /*
@@ -140,8 +140,7 @@ export class AST {
   getContainingRoot(node: ASTNode): SourceUnit {
     if (node instanceof SourceUnit) return node;
 
-    const root = node.getClosestParentByType(SourceUnit);
-    assert(root !== undefined, `Could not find root source unit for ${printNode(node)}`);
+    const root = getContainingSourceUnit(node);
     assert(
       this.roots.includes(root),
       `Found ${printNode(root)} as root of ${printNode(node)}, but this is not in the AST roots`,
@@ -185,11 +184,7 @@ export class AST {
   }
 
   getUtilFuncGen(node: ASTNode): CairoUtilFuncGen {
-    const sourceUnit = node instanceof SourceUnit ? node : node.getClosestParentByType(SourceUnit);
-    assert(
-      sourceUnit !== undefined,
-      'Could not find the sourceUnit to attach the nodes generated functions to',
-    );
+    const sourceUnit = node instanceof SourceUnit ? node : getContainingSourceUnit(node);
     const gen = this.cairoUtilFuncGen.get(sourceUnit.id);
     if (gen === undefined) {
       const newGen = new CairoUtilFuncGen(this, sourceUnit);
