+++ src/passes/expressionSplitter/conditionalFunctionaliser.ts
@@ -1,4 +1,3 @@
-import assert = require('assert');
 import {
   Assignment,
   ASTNode,
@@ -8,7 +7,6 @@ import {
   Expression,
   ExpressionStatement,
   FunctionCall,
-  FunctionDefinition,
   getNodeType,
   Identifier,
   IfStatement,
@@ -19,7 +17,6 @@ import {
   VariableDeclarationStatement,
 } from 'solc-typed-ast';
 import { AST } from '../../ast/ast';
-import { printNode } from '../../utils/astPrinter';
 import { cloneASTNode } from '../../utils/cloning';
 import { getDefaultValue } from '../../utils/defaultValueNodes';
 import { collectUnboundVariables, createOuterCall } from '../../utils/functionGeneration';
@@ -64,12 +61,6 @@ export function getConditionalReturnVariable(
     Mutability.Mutable,
     node.typeString,
   );
-}
-
-export function getContainingFunction(node: ASTNode): FunctionDefinition {
-  const func = node.getClosestParentByType(FunctionDefinition);
-  assert(func !== undefined, `Unable to find containing function for ${printNode(node)}`);
-  return func;
 }
 
 // The inputs to the function should be only the free variables
