+++ src/passes/loopFunctionaliser/breakToReturn.ts
@@ -1,17 +1,12 @@
-import assert from 'assert';
-import { Break, FunctionDefinition } from 'solc-typed-ast';
+import { Break } from 'solc-typed-ast';
 import { AST } from '../../ast/ast';
 import { ASTMapper } from '../../ast/mapper';
-import { printNode } from '../../utils/astPrinter';
 import { createReturn } from '../../utils/nodeTemplates';
+import { getContainingFunction } from '../../utils/utils';
 
 export class BreakToReturn extends ASTMapper {
   visitBreak(node: Break, ast: AST): void {
-    const containingFunction = node.getClosestParentByType(FunctionDefinition);
-    assert(
-      containingFunction !== undefined,
-      `Unable to find containing function for ${printNode(node)}`,
-    );
+    const containingFunction = getContainingFunction(node);
 
     ast.replaceNode(
       node,
