+++ src/utils/utils.ts
@@ -4,6 +4,7 @@ import {
   ArrayType,
   ArrayTypeName,
   Assignment,
+  ASTNode,
   BoolType,
   BytesType,
   CompileFailedError,
@@ -33,6 +34,7 @@ import {
   PointerType,
   Return,
   SourceLocation,
+  SourceUnit,
   StateVariableVisibility,
   StringType,
   StructDefinition,
@@ -47,7 +49,7 @@ import {
 import Web3 from 'web3';
 import { AST } from '../ast/ast';
 import { isSane } from './astChecking';
-import { printTypeNode } from './astPrinter';
+import { printNode, printTypeNode } from './astPrinter';
 import {
   logError,
   NotSupportedYetError,
@@ -479,4 +481,16 @@ export function getSourceFromLocation(source: string, location: SourceLocation):
     .reduce(([s, n], c) => [[...s, `${n}  ${c}`], n + 1], [new Array<string>(), followingLineNum]);
 
   return [...previousLines, ...currentLine, ...followingLines].join('\n');
+}
+
+export function getContainingFunction(node: ASTNode): FunctionDefinition {
+  const func = node.getClosestParentByType(FunctionDefinition);
+  assert(func !== undefined, `Unable to find containing function for ${printNode(node)}`);
+  return func;
+}
+
+export function getContainingSourceUnit(node: ASTNode): SourceUnit {
+  const root = node.getClosestParentByType(SourceUnit);
+  assert(root !== undefined, `Unable to find root source unit for ${printNode(node)}`);
+  return root;
 }
